# api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Bus Tracking API - v2 Dynamic",
        "description": "Refactored API supporting dynamic Trip orchestration between Drivers, Buses, and Routes.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/api/driver/start-trip": {
            "post": {
                "summary": "Start Trip (Driver)",
                "tags": [
                    "Driver"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "busId",
                                    "routeId"
                                ],
                                "properties": {
                                    "busId": {
                                        "type": "string"
                                    },
                                    "routeId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/driver/stop-trip": {
            "post": {
                "summary": "Stop Trip (Pause) (Driver)",
                "tags": [
                    "Driver"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "tripId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/driver/resume-trip": {
            "post": {
                "summary": "Resume Trip (Driver)",
                "tags": [
                    "Driver"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "tripId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/driver/end-trip": {
            "post": {
                "summary": "End Trip (Driver)",
                "tags": [
                    "Driver"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "tripId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}